Implementing all types and methods generated

+++ src/methods/get_user_profile_photos.rs
@@ -1,23 +1,22 @@
+/// This code is generated using teleborg-api-generator (https://gitlab.com/b.wisman155/teleborg-api-generator)
 use super::Method;
 use crate::types::UserProfilePhotos;
 use serde::Serialize;
 use typed_builder::TypedBuilder;
 
-/// Use this method to get a list of profile pictures for a user. Returns a `UserProfilePhotos`
-/// struct instance.
+/// Use this method to get a list of profile pictures for a user. Returns a UserProfilePhotos object.
 #[derive(Debug, TypedBuilder, Serialize)]
 pub struct GetUserProfilePhotos {
-    /// Unique identifier of the target user.
-    pub user_id: i32,
+    /// Unique identifier of the target user
+    pub user_id: i64,
     #[builder(default)]
     #[serde(skip_serializing_if = "Option::is_none")]
     /// Sequential number of the first photo to be returned. By default, all photos are returned.
-    pub offset: Option<i32>,
+    pub offset: Option<i64>,
     #[builder(default)]
     #[serde(skip_serializing_if = "Option::is_none")]
-    /// Limits the number of photos to be retrieved. Values between 1 - 100 are accepted. Defaults
-    /// to 100.
-    pub limit: Option<i32>,
+    /// Limits the number of photos to be retrieved. Values between 1—100 are accepted. Defaults to 100.
+    pub limit: Option<u8>,
 }
 
 impl_method!(
